+++ packages/sdk/browser/contract-tests/entity/src/ClientEntity.ts
@@ -29,7 +29,6 @@ function makeSdkConfig(options: SDKConfigParams, tag: string) {
   }
 
   if (options.polling) {
-    cf.initialConnectionMode = 'polling';
     if (options.polling.baseUri) {
       cf.baseUri = options.polling.baseUri;
     }
@@ -42,7 +41,7 @@ function makeSdkConfig(options: SDKConfigParams, tag: string) {
     if (options.streaming.baseUri) {
       cf.streamUri = options.streaming.baseUri;
     }
-    cf.initialConnectionMode = 'streaming';
+    cf.streaming = true;
     cf.streamInitialReconnectDelay = maybeTime(options.streaming.initialRetryDelayMs);
   }
 
@@ -157,9 +156,7 @@ export class ClientEntity {
         if (!identifyParams) {
           throw malformedCommand;
         }
-        await this.client.identify(identifyParams.user || identifyParams.context, {
-          waitForNetworkResults: true,
-        });
+        await this.client.identify(identifyParams.user || identifyParams.context);
         return undefined;
       }
 
@@ -205,7 +202,7 @@ export async function newSdkClientEntity(options: CreateInstanceParams) {
   let failed = false;
   try {
     await Promise.race([
-      client.identify(initialContext, { waitForNetworkResults: true }),
+      client.identify(initialContext),
       new Promise((_resolve, reject) => {
         setTimeout(reject, timeout);
       }),