+++ packages/shared/mocks/jest.config.js
@@ -1,7 +1,0 @@
-module.exports = {
-  transform: { '^.+\\.ts?$': 'ts-jest' },
-  testMatch: ['**/*.test.ts?(x)'],
-  testEnvironment: 'node',
-  moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx', 'json', 'node'],
-  collectCoverageFrom: ['src/**/*.ts'],
-};