+++ packages/shared/sdk-client/__tests__/LDClientImpl.events.test.ts
@@ -2,36 +2,39 @@ import {
   AutoEnvAttributes,
   ClientContext,
   clone,
-  Encoding,
   internal,
   LDContext,
+  LDLogger,
   subsystem,
 } from '@launchdarkly/js-sdk-common';
-import {
-  createBasicPlatform,
-  createLogger,
-  MockEventProcessor,
-} from '@launchdarkly/private-js-mocks';
 
 import LDClientImpl from '../src/LDClientImpl';
 import { Flags } from '../src/types';
+import { createBasicPlatform } from './createBasicPlatform';
 import * as mockResponseJson from './evaluation/mockResponse.json';
+import { MockEventProcessor } from './eventProcessor';
 import { MockEventSource } from './streaming/LDClientImpl.mocks';
+import { makeTestDataManagerFactory } from './TestDataManager';
 
 type InputCustomEvent = internal.InputCustomEvent;
 type InputIdentifyEvent = internal.InputIdentifyEvent;
 
 let mockPlatform: ReturnType<typeof createBasicPlatform>;
-let logger: ReturnType<typeof createLogger>;
+let logger: LDLogger;
 
 beforeEach(() => {
   mockPlatform = createBasicPlatform();
-  logger = createLogger();
+  logger = {
+    error: jest.fn(),
+    warn: jest.fn(),
+    info: jest.fn(),
+    debug: jest.fn(),
+  };
 });
 
 jest.mock('@launchdarkly/js-sdk-common', () => {
   const actual = jest.requireActual('@launchdarkly/js-sdk-common');
-  const m = jest.requireActual('@launchdarkly/private-js-mocks');
+  const m = jest.requireActual('./eventProcessor');
   return {
     ...actual,
     ...{
@@ -80,18 +83,15 @@ describe('sdk-client object', () => {
 
     mockPlatform.crypto.randomUUID.mockReturnValue('random1');
 
-    ldc = new LDClientImpl(testSdkKey, AutoEnvAttributes.Enabled, mockPlatform, {
-      logger,
-    });
-
-    jest.spyOn(LDClientImpl.prototype as any, 'getStreamingPaths').mockReturnValue({
-      pathGet(_encoding: Encoding, _plainContextString: string): string {
-        return '/stream/path';
+    ldc = new LDClientImpl(
+      testSdkKey,
+      AutoEnvAttributes.Enabled,
+      mockPlatform,
+      {
+        logger,
       },
-      pathReport(_encoding: Encoding, _plainContextString: string): string {
-        return '/stream/path';
-      },
-    });
+      makeTestDataManagerFactory(testSdkKey, mockPlatform),
+    );
   });
 
   afterEach(() => {