+++ packages/shared/sdk-server/__tests__/LDClient.hooks.test.ts
@@ -1,8 +1,7 @@
-import { createBasicPlatform } from '@launchdarkly/private-js-mocks';
-
 import { LDClientImpl, LDMigrationStage } from '../src';
 import Reasons from '../src/evaluation/Reasons';
 import TestData from '../src/integrations/test_data/TestData';
+import { createBasicPlatform } from './createBasicPlatform';
 import { TestHook } from './hooks/TestHook';
 import TestLogger from './Logger';
 import makeCallbacks from './makeCallbacks';
@@ -20,7 +19,7 @@ describe('given an LDClient with test data', () => {
     testHook = new TestHook();
     td = new TestData();
     client = new LDClientImpl(
-      'sdk-key',
+      'sdk-key-hooks-test-data',
       createBasicPlatform(),
       {
         updateProcessor: td.getFactory(),
@@ -351,7 +350,7 @@ it('can add a hook after initialization', async () => {
   const logger = new TestLogger();
   const td = new TestData();
   const client = new LDClientImpl(
-    'sdk-key',
+    'sdk-key-hook-after-init',
     createBasicPlatform(),
     {
       updateProcessor: td.getFactory(),